Defer kl_divergence computation to base distributions of two instances of the Transformed if they use the same bijector.
KL divergence is invariant under invertible transformations. If two distributions are transformed by the same bijector, the value of the KL divergence does not change. Therefore, instead of computing the KL divergence between two instances of Transformed, it is enough to compute the KL between their base distributions if they use the same bijector. This allows e.g. computing KL between two flows if they use MVN as the base and the same bijector (currently in such a case we get an error that there is no method to compute KL between the instances of Transformed).

PiperOrigin-RevId: 369101147