[JAX] Relax test tolerances in distrax multinomial_test.
An upcoming change to JAX switches to using MLIR when lowering JAX to XLA. This causes a small change to the numerics because we end up constant folding an expression that XLA would not have previously constant-folded.

In principle this should cause a *more* accurate result rather than a less-accurate one, since it means we apply the following rewrite inside a constant evaluation of an `lgamma` operation:

```
-  tmp_45 = f32[] constant(1.50563267e-07)
-  tmp_46 = f32[2]{0} broadcast(tmp_45), dimensions={}
-  tmp_47 = f32[2]{0} constant({11, 18})
-  tmp_48 = f32[2]{0} divide(tmp_46, tmp_47)
...
+  tmp_45 = f32[2]{0} constant({1.36875693e-08, 8.36462632e-09})
```
i.e., performing a division at compile time on CPU. On GPU and TPU XLA will lower the same computation to multiplication by the reciprocal, which is less accurate. However for some reason that I haven't fully tracked down a *more* accurate division leads to a *less* accurate output for these test cases..

Relax test tolerances to keep the test passing.

PiperOrigin-RevId: 416326460