Implement TFP's base measure interface for distrax dists and bijectors.
The interface needs two new methods: an `experimental_local_measure` on `Distribution`s,
and an `experimental_density_correction` on `Bijector`s. We implement these methods
using the default assumptions already made by TFP: that the tangent space to a manifold
has the standard basis on R^n, and that the transformations on this manifold are
dimension-preserving. In the future, TFP will define more appropriate bases and
density corrections for subclasses of `Distribution` and `Bijector`; distrax is not expected
to follow this protocol.

See:
- `tfp.python.experimental.tangent_spaces`
- `tfp.distributions.experimental_local_measure`
- `tfp.bijectors.experimental_compute_density_correction`
- Radul and Alexeev, AISTATS 2021, “The Base Measure Problem and its Solution”, https://arxiv.org/abs/2010.09647.
PiperOrigin-RevId: 431515351