Add survival and log-survival function to `distrax.Distribution` base class.
Note that by default we use a numerically not necessarily stable definition
of the survival function in terms of the CDF: sf(x) = 1-cdf(x) and similarly
for the log survival function.
More stable definitions should be implemented in future CLs for distributions,
for which they exist.

Added a test to `distribution.Uniform` where we will use the default behaviour.

PiperOrigin-RevId: 443652231