Add PyPI badge showing the latest released version.
Recommend installing the PyPI released package first, instead of installing from
GitHub, as it may not be as stable as the released version.

Fix the GitHub URL to use `git+https://github.com/...` instead of `git+git://...`.

Use the block code format for `pip` installation commands instead of inline code
for improved readability and easier copy-pasting.

PiperOrigin-RevId: 463148703