Ignore bad return type annotation in convert_seed_and_sample_shape
Why? This annotation is incorrect (the type checker does not consider `rng` here to be a `PRNGKey`) The reason this has worked in the past is because `chex.PRNGKey` has been effectively considered to be `Any`; this incorrect annotation is blocking us from making `chex.PRNGKey` a more meaningful annotation.

PiperOrigin-RevId: 565184594